Update Vote method

+++ plugins/poll/lib/poll.rb
@@ -9,7 +9,12 @@ class DiscoursePoll::Poll
         poll_id = poll.id
         # remove options that aren't available in the poll
         available_options = poll.poll_options.map { |o| o.digest }.to_set
-        options.select! { |o| available_options.include?(o) }
+
+        if poll.type == "irv"
+          options.select! { |o| available_options.include?(o[:digest]) }
+        else
+          options.select! { |o| available_options.include?(o) }
+        end
 
         if options.empty?
           raise DiscoursePoll::Error.new I18n.t("poll.requires_at_least_1_valid_option")
@@ -19,7 +24,11 @@ class DiscoursePoll::Poll
           poll
             .poll_options
             .each_with_object([]) do |option, obj|
-              obj << option.id if options.include?(option.digest)
+              if poll.type == "irv"
+                obj << option.id if options.any? { |o| o[:digest] == option.digest }
+              else
+                obj << option.id if options.include?(option.digest)
+              end
             end
 
         self.validate_votes!(poll, new_option_ids)
@@ -36,30 +45,42 @@ class DiscoursePoll::Poll
 
         # create missing votes
         (new_option_ids - old_option_ids).each do |option_id|
-          PollVote.create!(poll: poll, user: user, poll_option_id: option_id)
+          if poll.type == "irv"
+            PollVote.create!(
+              poll: poll,
+              user: user,
+              poll_option_id: option_id,
+              rank:
+                options.find { |o| o[:digest] == poll.poll_options.find(option_id).digest }[:rank],
+            )
+          else
+            PollVote.create!(poll: poll, user: user, poll_option_id: option_id)
+          end
         end
       end
 
-    # Ensure consistency here as we do not have a unique index to limit the
-    # number of votes per the poll's configuration.
-    is_multiple = serialized_poll[:type] == "multiple"
-    offset = is_multiple ? (serialized_poll[:max] || serialized_poll[:options].length) : 1
+    unless serialized_poll[:type] == "irv"
+      # Ensure consistency here as we do not have a unique index to limit the
+      # number of votes per the poll's configuration.
+      is_multiple = serialized_poll[:type] == "multiple"
+      offset = is_multiple ? (serialized_poll[:max] || serialized_poll[:options].length) : 1
 
-    DB.query(<<~SQL, poll_id: poll_id, user_id: user.id, offset: offset)
-    DELETE FROM poll_votes
-    USING (
-      SELECT
-        poll_id,
-        user_id
-      FROM poll_votes
-      WHERE poll_id = :poll_id
-      AND user_id = :user_id
-      ORDER BY created_at DESC
-      OFFSET :offset
-    ) to_delete_poll_votes
-    WHERE poll_votes.poll_id = to_delete_poll_votes.poll_id
-    AND poll_votes.user_id = to_delete_poll_votes.user_id
-    SQL
+      DB.query(<<~SQL, poll_id: poll_id, user_id: user.id, offset: offset)
+      DELETE FROM poll_votes
+      USING (
+        SELECT
+          poll_id,
+          user_id
+        FROM poll_votes
+        WHERE poll_id = :poll_id
+        AND user_id = :user_id
+        ORDER BY created_at DESC
+        OFFSET :offset
+      ) to_delete_poll_votes
+      WHERE poll_votes.poll_id = to_delete_poll_votes.poll_id
+      AND poll_votes.user_id = to_delete_poll_votes.user_id
+      SQL
+    end
 
     [serialized_poll, options]
   end
@@ -362,6 +383,16 @@ class DiscoursePoll::Poll
         raise DiscoursePoll::Error.new(I18n.t("poll.min_vote_per_user", count: poll.min))
       elsif poll.max && (num_of_options > poll.max)
         raise DiscoursePoll::Error.new(I18n.t("poll.max_vote_per_user", count: poll.max))
+      end
+    elsif poll.irv?
+      if poll.poll_options.length != num_of_options
+        raise DiscoursePoll::Error.new(
+                I18n.t(
+                  "poll.irv_vote_options_mismatch",
+                  count: poll.options.length,
+                  provided: num_of_options,
+                ),
+              )
       end
     elsif num_of_options > 1
       raise DiscoursePoll::Error.new(I18n.t("poll.one_vote_per_user"))
